fix not logged in redirection, localstorage jwt

+++ src/components/auth/Logout.jsx
@@ -1,35 +1,32 @@
 import React, { useState, useEffect } from "react";
 import { Navigate } from "react-router-dom";
+import { URL_API } from "../../config";
+import { checkLogin } from "../../utils/checkLogin";
+import { fetchApi } from "../../utils/response";
 
-export default function Logout(props) {
+export default function Logout({ props }) {
   const [hasLoggedOut, setHasLoggedOut] = useState(false);
-
-  const [isLogged, setIsLogged] = useState(false);
+  const [isLogged, setIsLogged] = useState(true);
 
   useEffect(() => {
-    localStorage.getItem("token") == null
-      ? setIsLogged(false)
-      : setIsLogged(true);
-  }, []);
+    setIsLogged(props.isLogged);
+  }, [props.isLogged]);
 
-  function logout(e) {
+  async function logout(e) {
     e.preventDefault();
 
+    const response = await fetchApi(URL_API + "/auth/logout", "POST");
+
+    if (!response.ok) return console.log(response.data.status);
+
     localStorage.clear();
-
     setHasLoggedOut(true);
-  }
-
-  function checkLog() {
-    console.log(isLogged);
-    console.log(hasLoggedOut);
+    props.setIsLogged(false);
   }
 
   return (
     <>
       <h2>Logout</h2>
-      {checkLog()}
-
       {!isLogged ? (
         <Navigate to="/auth/login" replace />
       ) : !hasLoggedOut ? (
